// src/app/api/model/badgeResponse.ts
/**
 * API EDUGO - Plateforme Éducative Malienne
 *
 * Contact: support@edugo.ml
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BadgeResponse { 
    id?: number;
    nom?: string;
    description?: string;
    type?: string;
    icone?: string;
}