// src/app/api/model/statistiquesPlateformeResponse.ts
/**
 * API EDUGO - Plateforme Éducative Malienne
 *
 * Contact: support@edugo.ml
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { StatistiquesClasseResponse } from './statistiquesClasseResponse';
import { StatistiquesNiveauResponse } from './statistiquesNiveauResponse';


export interface StatistiquesPlateformeResponse { 
    totalUtilisateurs?: number;
    utilisateursActifs?: number;
    utilisateursInactifs?: number;
    totalEleves?: number;
    totalAdmins?: number;
    totalLivres?: number;
    livresDisponibles?: number;
    livresIndisponibles?: number;
    totalExercices?: number;
    exercicesActifs?: number;
    exercicesInactifs?: number;
    totalQuiz?: number;
    quizActifs?: number;
    quizInactifs?: number;
    totalDefis?: number;
    defisActifs?: number;
    defisInactifs?: number;
    totalChallenges?: number;
    challengesActifs?: number;
    challengesInactifs?: number;
    totalExercicesRealises?: number;
    totalQuizCompletes?: number;
    totalDefisReussis?: number;
    totalChallengesReussis?: number;
    totalPointsGagnes?: number;
    totalBadges?: number;
    totalBadgesAttribues?: number;
    tauxCompletionMoyen?: number;
    totalProgressions?: number;
    dateDerniereMiseAJour?: string;
    dateDebutPeriode?: string;
    dateFinPeriode?: string;
    statistiquesParNiveau?: Array<StatistiquesNiveauResponse>;
    statistiquesParClasse?: Array<StatistiquesClasseResponse>;
}